refine, code refine for oracle server.
