+++ oracle_server/oracle_server_test.go
@@ -7,13 +7,7 @@ import (
 	"autonity-oracle/helpers"
 	"autonity-oracle/types"
 	"autonity-oracle/types/mock"
-	"fmt"
-	"io/ioutil" //nolint
-	"math"
 	"math/big"
-	"os"
-	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 
@@ -192,7 +186,7 @@ func TestOracleServer(t *testing.T) {
 			}
 		}
 
-		voteRecord, err := srv.assembleReportData(srv.curRound, helpers.DefaultSymbols, prices)
+		voteRecord, err := srv.assembleVote(srv.curRound, helpers.DefaultSymbols, prices)
 		require.NoError(t, err)
 		srv.voteRecords[srv.curRound] = voteRecord
 
@@ -258,153 +252,6 @@ func TestOracleServer(t *testing.T) {
 		srv.runningPlugins["template_plugin"].Close()
 	})
 
-	t.Run("test plugin runtime management, add new plugin", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		dialerMock := mock.NewMockDialer(ctrl)
-		contractMock := cMock.NewMockContractAPI(ctrl)
-		contractMock.EXPECT().GetRound(nil).Return(currentRound, nil)
-		contractMock.EXPECT().GetSymbols(nil).Return(helpers.DefaultSymbols, nil)
-		contractMock.EXPECT().GetVotePeriod(nil).Return(votePeriod, nil)
-		contractMock.EXPECT().WatchNewRound(gomock.Any(), gomock.Any()).Return(subRoundEvent, nil)
-		contractMock.EXPECT().WatchNewSymbols(gomock.Any(), gomock.Any()).Return(subSymbolsEvent, nil)
-		contractMock.EXPECT().WatchPenalized(gomock.Any(), gomock.Any(), gomock.Any()).Return(subPenalizeEvent, nil)
-		contractMock.EXPECT().WatchSuccessfulVote(gomock.Any(), gomock.Any(), gomock.Any()).Return(subVoteEvent, nil)
-		contractMock.EXPECT().WatchTotalOracleRewards(gomock.Any(), gomock.Any()).Return(subReportedEvent, nil)
-		contractMock.EXPECT().WatchInvalidVote(gomock.Any(), gomock.Any(), gomock.Any()).Return(subInvalidVoteEvent, nil)
-		l1Mock := mock.NewMockBlockchain(ctrl)
-		l1Mock.EXPECT().ChainID(gomock.Any()).Return(ChainIDPiccadilly, nil)
-		srv := NewOracleServer(conf, dialerMock, l1Mock, contractMock)
-		require.Equal(t, currentRound.Uint64(), srv.curRound)
-		require.Equal(t, DefaultSampledSymbols, srv.samplingSymbols)
-		require.Equal(t, true, srv.pricePrecision.Equal(decimal.NewFromBigInt(common.Big1, int32(precision))))
-
-		require.Equal(t, votePeriod.Uint64(), srv.votePeriod)
-		require.Equal(t, 1, len(srv.runningPlugins))
-
-		// add a new plugin into the plugin directory.
-		clones, err := clonePlugins(srv.conf.PluginDIR, "cloned", srv.conf.PluginDIR)
-		require.NoError(t, err)
-		defer func() {
-			for _, f := range clones {
-				err := os.Remove(f)
-				require.NoError(t, err)
-			}
-		}()
-
-		srv.PluginRuntimeManagement()
-		require.Equal(t, 2, len(srv.runningPlugins))
-		require.Equal(t, "template_plugin", srv.runningPlugins["template_plugin"].Name())
-		require.Equal(t, "clonedtemplate_plugin", srv.runningPlugins["clonedtemplate_plugin"].Name())
-		srv.runningPlugins["template_plugin"].Close()
-		srv.runningPlugins["clonedtemplate_plugin"].Close()
-	})
-
-	t.Run("test plugin runtime management, upgrade plugin", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		dialerMock := mock.NewMockDialer(ctrl)
-		contractMock := cMock.NewMockContractAPI(ctrl)
-		contractMock.EXPECT().GetRound(nil).Return(currentRound, nil)
-		contractMock.EXPECT().GetSymbols(nil).Return(helpers.DefaultSymbols, nil)
-		contractMock.EXPECT().GetVotePeriod(nil).Return(votePeriod, nil)
-		contractMock.EXPECT().WatchNewRound(gomock.Any(), gomock.Any()).Return(subRoundEvent, nil)
-		contractMock.EXPECT().WatchNewSymbols(gomock.Any(), gomock.Any()).Return(subSymbolsEvent, nil)
-		contractMock.EXPECT().WatchPenalized(gomock.Any(), gomock.Any(), gomock.Any()).Return(subPenalizeEvent, nil)
-		contractMock.EXPECT().WatchSuccessfulVote(gomock.Any(), gomock.Any(), gomock.Any()).Return(subVoteEvent, nil)
-		contractMock.EXPECT().WatchTotalOracleRewards(gomock.Any(), gomock.Any()).Return(subReportedEvent, nil)
-		contractMock.EXPECT().WatchInvalidVote(gomock.Any(), gomock.Any(), gomock.Any()).Return(subInvalidVoteEvent, nil)
-		l1Mock := mock.NewMockBlockchain(ctrl)
-		l1Mock.EXPECT().ChainID(gomock.Any()).Return(ChainIDPiccadilly, nil)
-		srv := NewOracleServer(conf, dialerMock, l1Mock, contractMock)
-		require.Equal(t, currentRound.Uint64(), srv.curRound)
-		require.Equal(t, DefaultSampledSymbols, srv.samplingSymbols)
-		require.Equal(t, true, srv.pricePrecision.Equal(decimal.NewFromBigInt(common.Big1, int32(precision))))
-
-		require.Equal(t, votePeriod.Uint64(), srv.votePeriod)
-		require.Equal(t, 1, len(srv.runningPlugins))
-		firstStart := srv.runningPlugins["template_plugin"].StartTime()
-
-		// cpy and replace the legacy plugins
-		err := replacePlugins(srv.conf.PluginDIR)
-		require.NoError(t, err)
-
-		srv.PluginRuntimeManagement()
-
-		require.Equal(t, 1, len(srv.runningPlugins))
-		require.Equal(t, "template_plugin", srv.runningPlugins["template_plugin"].Name())
-		require.Greater(t, srv.runningPlugins["template_plugin"].StartTime(), firstStart)
-		srv.runningPlugins["template_plugin"].Close()
-	})
-
-	t.Run("test plugin runtime management, remove plugin", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		dialerMock := mock.NewMockDialer(ctrl)
-		contractMock := cMock.NewMockContractAPI(ctrl)
-		contractMock.EXPECT().GetRound(nil).Return(currentRound, nil)
-		contractMock.EXPECT().GetSymbols(nil).Return(helpers.DefaultSymbols, nil)
-		contractMock.EXPECT().GetVotePeriod(nil).Return(votePeriod, nil)
-		contractMock.EXPECT().WatchNewRound(gomock.Any(), gomock.Any()).Return(subRoundEvent, nil)
-		contractMock.EXPECT().WatchNewSymbols(gomock.Any(), gomock.Any()).Return(subSymbolsEvent, nil)
-		contractMock.EXPECT().WatchPenalized(gomock.Any(), gomock.Any(), gomock.Any()).Return(subPenalizeEvent, nil)
-		contractMock.EXPECT().WatchSuccessfulVote(gomock.Any(), gomock.Any(), gomock.Any()).Return(subVoteEvent, nil)
-		contractMock.EXPECT().WatchTotalOracleRewards(gomock.Any(), gomock.Any()).Return(subReportedEvent, nil)
-		contractMock.EXPECT().WatchInvalidVote(gomock.Any(), gomock.Any(), gomock.Any()).Return(subInvalidVoteEvent, nil)
-
-		l1Mock := mock.NewMockBlockchain(ctrl)
-		l1Mock.EXPECT().ChainID(gomock.Any()).Return(ChainIDPiccadilly, nil)
-		srv := NewOracleServer(conf, dialerMock, l1Mock, contractMock)
-		require.Equal(t, currentRound.Uint64(), srv.curRound)
-		require.Equal(t, DefaultSampledSymbols, srv.samplingSymbols)
-		require.Equal(t, true, srv.pricePrecision.Equal(decimal.NewFromBigInt(common.Big1, int32(precision))))
-
-		require.Equal(t, votePeriod.Uint64(), srv.votePeriod)
-		require.Equal(t, 1, len(srv.runningPlugins))
-
-		// Backup the existing plugins
-		backupDir := "/tmp/plugin_backup"   // Temporary directory for backup
-		err := os.MkdirAll(backupDir, 0755) // Create backup directory if it doesn't exist
-		require.NoError(t, err)
-
-		// Copy existing plugins to the backup directory
-		files, err := ioutil.ReadDir(srv.conf.PluginDIR)
-		require.NoError(t, err)
-		for _, file := range files {
-			if !file.IsDir() && helpers.IsExecOwnerGroup(file.Mode()) {
-				src := filepath.Join(srv.conf.PluginDIR, file.Name())
-				dst := filepath.Join(backupDir, file.Name())
-				err = os.Link(src, dst) // Use os.Link for hard link; use os.Copy for actual copy if needed
-				require.NoError(t, err)
-			}
-		}
-
-		// Defer the recovery action to restore the removed plugins
-		defer func() {
-			// Restore the plugins from the backup directory
-			files, err = ioutil.ReadDir(backupDir)
-			require.NoError(t, err)
-			for _, file := range files {
-				src := filepath.Join(backupDir, file.Name())
-				dst := filepath.Join(srv.conf.PluginDIR, file.Name())
-				err = os.Link(src, dst) // Use os.Link for hard link; use os.Copy for actual copy if needed
-				require.NoError(t, err)
-			}
-			os.RemoveAll(backupDir) // Clean up the backup directory
-		}()
-
-		// Remove the plugins
-		err = removePlugins(srv.conf.PluginDIR)
-		require.NoError(t, err)
-
-		srv.PluginRuntimeManagement()
-
-		require.Equal(t, 0, len(srv.runningPlugins))
-	})
-
 	t.Run("gcRounddata", func(t *testing.T) {
 		os := &OracleServer{
 			voteRecords: make(map[uint64]*types.VoteRecord),
@@ -421,131 +268,4 @@ func TestOracleServer(t *testing.T) {
 		require.Equal(t, MaxBufferedRounds, len(os.voteRecords))
 
 	})
-}
-
-// clone plugins from a src directory to new directory by adding prefix in the name of each binary, and return the cloned
-// new file names and an error.
-func clonePlugins(pluginDIR string, clonePrefix string, destDir string) ([]string, error) {
-
-	var clonedPlugins []string
-	files, err := helpers.ListPlugins(pluginDIR)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, file := range files {
-		// read srcFile
-		srcContent, err := ioutil.ReadFile(fmt.Sprintf("%s/%s", pluginDIR, file.Name()))
-		if err != nil {
-			return clonedPlugins, err
-		}
-
-		// create dstFile and copy the content
-		newPlugin := fmt.Sprintf("%s/%s%s", destDir, clonePrefix, file.Name())
-		err = ioutil.WriteFile(newPlugin, srcContent, file.Mode())
-		if err != nil {
-			return clonedPlugins, err
-		}
-		clonedPlugins = append(clonedPlugins, newPlugin)
-	}
-	return clonedPlugins, nil
-}
-
-func replacePlugins(pluginDir string) error {
-	rawPlugins, err := helpers.ListPlugins(pluginDir)
-	if err != nil {
-		return err
-	}
-
-	clonePrefix := "clone"
-	clonedPlugins, err := clonePlugins(pluginDir, clonePrefix, fmt.Sprintf("%s/..", pluginDir))
-	if err != nil {
-		return err
-	}
-
-	for _, file := range clonedPlugins {
-		for _, info := range rawPlugins {
-			if strings.Contains(file, info.Name()) {
-				err := os.Rename(file, fmt.Sprintf("%s/%s", pluginDir, info.Name()))
-				if err != nil {
-					return err
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-// removePlugins removes all executable plugin binaries from the specified plugin directory.
-func removePlugins(pluginDir string) error {
-	// Read the directory
-	files, err := ioutil.ReadDir(pluginDir)
-	if err != nil {
-		return err
-	}
-
-	// Iterate over the files
-	for _, file := range files {
-		// Skip directories
-		if file.IsDir() {
-			continue
-		}
-
-		// Check if the file is an executable
-		if helpers.IsExecOwnerGroup(file.Mode()) {
-			// Construct the full file path
-			filePath := pluginDir + "/" + file.Name()
-			// Remove the file
-			if err := os.Remove(filePath); err != nil {
-				return err // Return error if unable to remove the file
-			}
-		}
-	}
-	return nil
-}
-
-// TestComputeConfidence tests the ComputeConfidence function.
-func TestComputeConfidence(t *testing.T) {
-	tests := []struct {
-		symbol       string
-		numOfSamples int
-		strategy     int
-		expected     uint8
-	}{
-		// Forex symbols with ConfidenceStrategyFixed, max confidence are expected.
-		{"AUD-USD", 1, config.ConfidenceStrategyFixed, MaxConfidence},
-		{"CAD-USD", 2, config.ConfidenceStrategyFixed, MaxConfidence},
-		{"EUR-USD", 3, config.ConfidenceStrategyFixed, MaxConfidence},
-		{"GBP-USD", 4, config.ConfidenceStrategyFixed, MaxConfidence},
-		{"JPY-USD", 5, config.ConfidenceStrategyFixed, MaxConfidence},
-		{"SEK-USD", 10, config.ConfidenceStrategyFixed, MaxConfidence},
-
-		// Forex symbols with ConfidenceStrategyLinear
-		{"AUD-USD", 1, config.ConfidenceStrategyLinear, uint8(BaseConfidence + SourceScalingFactor*uint64(math.Pow(1.75, 1)))}, //nolint
-		{"CAD-USD", 2, config.ConfidenceStrategyLinear, uint8(BaseConfidence + SourceScalingFactor*uint64(math.Pow(1.75, 2)))}, //nolint
-		{"EUR-USD", 3, config.ConfidenceStrategyLinear, uint8(BaseConfidence + SourceScalingFactor*uint64(math.Pow(1.75, 3)))}, //nolint
-		{"GBP-USD", 4, config.ConfidenceStrategyLinear, MaxConfidence},
-		{"JPY-USD", 5, config.ConfidenceStrategyLinear, MaxConfidence},
-		{"SEK-USD", 10, config.ConfidenceStrategyLinear, MaxConfidence},
-
-		// Non-forex symbols with ConfidenceStrategyLinear, max confidence are expected.
-		{"ATN-USD", 1, config.ConfidenceStrategyLinear, MaxConfidence},
-		{"NTN-USD", 1, config.ConfidenceStrategyLinear, MaxConfidence},
-		{"NTN-ATN", 1, config.ConfidenceStrategyLinear, MaxConfidence},
-
-		// Non-forex symbols with ConfidenceStrategyFixed, max confidence are expected.
-		{"ATN-USD", 1, config.ConfidenceStrategyFixed, MaxConfidence},
-		{"NTN-USD", 1, config.ConfidenceStrategyFixed, MaxConfidence},
-		{"NTN-ATN", 1, config.ConfidenceStrategyFixed, MaxConfidence},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.symbol, func(t *testing.T) {
-			got := ComputeConfidence(tt.symbol, tt.numOfSamples, tt.strategy)
-			if got != tt.expected {
-				t.Errorf("ComputeConfidence(%q, %d, %d) = %d; want %d", tt.symbol, tt.numOfSamples, tt.strategy, got, tt.expected)
-			}
-		})
-	}
 }
